Added GTests

+++ lab1/include/Solution.hpp
@@ -5,8 +5,8 @@ namespace lab1{
 
     class Solution{
     public:
-        static std::pair<int, int> closest_pair_tonum(int upper_limit);
-        static double sqrt(int x);
+        static std::pair<int, int> closest_pair_tonum(int64_t upper_limit);
+        static double sqrt(int64_t  x);
         static double abs(double x);
     };
